This reorganizes the templates and bootenvs a little.
Sub-templating is used to get some code reuse out of the
kickstart and preseed files.  There are 5 or 6 new
subtemplates and example profiles to drive them.

An upgrade doc has been added that will collect info
about what to do as part of upgrading.

DO NOT USE THE join-to-dr.tmpl.  It will explode.