This is the completion of work started by @niemanme.
This adds as set of CLI helpers to set various subnet fields without
having to go through the json blob paths.

This also fixes a latent bug in the assureDecode code path.
GIN was sending 400 with strings and confusing the client on
bad json object.  This needs to use a lower level call to avoid
the GIN help.