For Content objects, add some additional meta data pieces to track usage.
Writable = true/false
Overwritable = true/false
Type = writable, local, dynamic, default

Writable is if the content's objects can be modified.
Overwritable is if the contents objects can be overwritten by
  other content layers.
Type is what layer the content is place.
  writable is highest layer (where objects are stored)
  local is the next read-only layer that is used for local changes
  dynamic is a set of uploadable content that is read-only (API content objects)
  default is the lowest layer that can be overwritten

This patch also makes the soft errors of validation returned in
Create or Update operations as a list in Warnings.