Add beta support for improved package repository handling.
This pull request adds support for improved package repository handling,
with an eye towards making it easy to configure for sites that use local
mirrors of common Linux distributions.

The major add is support in the core for a new param named
"package-repositories", along with several template helpers that know how
to render its contents appropriately.

The actual param is currently in the process of being added to tip for
the drp-community-content.  The code in this pull request will fall back to
current behaviour for OS installation if it is not defined in the system.

Initial pass at adding Render helpers for repository handling

Add fake debianish install test

Actually add the repo test file

Only try to rewrite package metadata if there is some.

Have explode_iso.sh only try to rewrite Yum repo package metadata
if there is metadata to rewrite, instead of just assuming it is always
there.

Repo rendering checkpoint for no-repo case

All unit tests pass with and without package-repositories defined

Local installs pass with new repo code