Fix busted Stage revalidation on Bootenv validation update
Stages were not properly clearing their validation state before
revalidating when the availability of their associated BootEnv was
updated.

This change also updates how Stages are validated to allow them to be
flagged as valid (and not available) as long as their are no obvious
syntactic issues with the Stage.  In particular, we no longer require
that the associated BootEnv, Task, Template, or Profiles actually
exist, although we do theck to make sure that the Template list is at
least structurally correct.