Add atomic operation capability to most actions.
The --ref flag can take a file or json or yaml blob
to test against before the patch is applied these
are used to build a full patch for update operations.

For the params command, --ref specifies the complete
set of parameters to test against.  It must match
completely before replacing with the new complete set.

For the set, add, and remove parameter operations,
the ref is the value of the parameter to test against.

set greg to fred --ref john

Only set greg to fred if the current value is john.

remove greg --ref john

Only remove greg if the current value is john.

Add is inheritly atomic.  Only add if not currently in
the parameter structure.  It does not use --ref.

If the --ref flag is left off, the operations work as before.