Tweak address cache for more proper operation.
First, logs have been threaded through the address cache, so we can
see what is being added to the local -> remote address cache at either
Debug or Trace log levels.

Second, close a potential hole that woul lead to deleting cache
entries too early by moving to a timeout-per-cache-line system instead
of an expired/non-expired bit, and increase the timeout to 10 minutes.

Third, close a potential race with what the kernel thinks each
connections local IP address should be by not asking for the local IP
address until the http connection state is Active, which means that
the client has sent us at least 1 byte.  The Linux kernel apparently
does not assign the local address until the handshake as completed,
and we were asking for the address before we checked to see that the
conn state was Active.