Option 67 is optional now.
You no longer have to specify option 67 in a subnet or a reservation,
the DHCP server now has sane defaults for what to do if it is missing
for legacy BIOS and UEFI systems, and for iPXE vs non-iPXE.  You can
still specify option 67 if you want to override the defaults.

If (for whatever reason), you do not want systems to PXE boot from a
subnet ever, you can toggle the new Unmanaged bit on the subnet.  This
will tell dr-provision to never send any PXE related options in a DHCP
response packet.