Have requestTracker.Find always return a copy.
Only the backend should be able to mess with objects in the indexes
directly.

To start fixing this up, arrange for Find to always return a copy of
the backend object instead of the object itself, unless we are calling
it from the backend directly.

Also add a test to catch the use case that is a consequence of not
having this check in place.