In order to assist the UX in figuring out subnet configuration, the interface object now has three new fields.
DnsServers - a list of strings that are the DnsServers configured on the
  DRP Endpoint.  This can be used to help Subnet definition.
DnsDomain - a string that represents that domain in the /etc/resolv.conf
  file or the host's domain if that isn't specified.
Gateway - The gateway address for the interface if specified.  This way
  the default gateway can be set for subnets that we know are using that
  gateway.

The UX is updated to populate default options with these values if
present.

This PR also fixes a bug in the bundler command that would take invalid
yaml files and create empty objects.