edit DR Provision license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit replaces the contents of the LICENSE file with the full text
of the Apache 2.0 license. It also moves the reference to the
documentation's CC licensing to the README.

These changes allow Licensee to successfully identify the license
type of DR Provision's codebase as Apache 2.0.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>