Add tenants.
Roles and Claims restrict what you can do, Tenants restrict what you
can see.

This begins the process of adding tenants and creating roles that
apply to the User making the request.  The strategy to do this involves
exposing a mostly-compiled set of Roles and Claims to the frontend
logic responsible for performing authentication, which is handled
by the models.Claims type.