Start working on scoped reservations.
This commit adds a new Scoped flag to Reservations, which signals that
the Reservation so created is only valid within a Subnet that has a
matching covered address range.  It will allow the DHCP system to
properly handle the case where we have multiple Reservations in
different Subnets for the same Token, which can happen when tagged
VLAN interfaces are in play on the client.

This pull request only implements the model and backend validation
changes, along with an update to the Subnet logic to disallow changes
to the Subnet field.  The changes to the DHCP system will come later.