update drpmachines.py to skip empty machines
This change will skip adding a machine to the inventory if the
machine name or the ip adddress is empty. This keeps context
runners out of the inventory, as well as other pseudo machines.

This change also introduces the ability to define a param named

    anisble_user

If set then that value will be used to override both root default,
and the shell env.

Signed-off-by: Michael Rice <michael@michaelrice.org>