Fixing error in ansible dynamic inventory
This change introduces a fix to allow the hostvar settings to be applied using the
ansible/hostvars param. Settings previously added to this param were not mapped.

This also fixes an error introduced in the previous commit which allowed the user to be set.
The user was never reset on the next loop iteration making an expected value be used.

Signed-off-by: Michael Rice <michael@michaelrice.org>