feat(fingerprint): machine fingerprint models and Gohai updates
Add the required models and machine code changes to allow for detcting
what Machine record most closely matches the current running machine.

Currently, this is based on a combination of matching certian classes of
DMI information, memory serial numbers, and MAC address matching.

This will eventually allow dr-provision to be resilient in the face
of hardware changes when running tasks.