feat(pool): Add pooling support
This commit adds the pooling objects and cli to the core
product.

The CLI and DOCs describe the operational and architectural usage.

Pools are optionally defined groups of machines that track
transition into and out the pools along with allocation and release.

Pools can define actions (add/remove parameters, add/remove profiles,
and change workflow) for entry and exit of pools and allocation and
release within the pool

Pool operations can be done on specific machines or filters that let
the user operate with cloud-like methods.