docs: Add example configuration for Hyper-V environment
Adds detailed description on how to build a DRP environment
targeted at Hyper-V for Windows 10 Pro or Enterprise. Can be
adapted to work with Hyper-V Hypervisor as well. Builds a complete
lab environment with a target machine and a DRP Endpoint using
the stable branch installer. The endpoint runs on a Debain 10
Virtual Machine, and one or more target Virtual Machines. The
target instaces are network isolated to allow for reproduction
of other environments with conflicting IP space.