feat(override-bootenvs): Add basic bootenv override functions.
We need a mechanism to allow for dynamic override of bootenvs to reign
in the bootenv explosions being caused by ESXi and others, and to
allow for user customizations of bootenvs that do not result in having
to copy everything that uses those bootenvs.

Along the way, switch to using deepcopy instead of JSON for model
cloning purposes.