fix(race): Fix client panic on multiple Close()
The client can panic of Close() is called on it multiple times, which
can be the case if several independent routines are using the same api
CLient and they both close the Client as part of cleanup.  Fix the
issue be refactoring the code to use a Context for closing instead.

While we are at it, have the agent refresh its token on a regular
schedule.