doc(faq): added place holder doc for drpy config
When someone manually removes a machine from drp after added it and having an
successful ESXi install done DRPY would flood the DRP endpoint with requests
causing drp cpu to spike to 100%. This doc will outline hwo to fix the issue.
For now its just a place holder so the code can have a faq number to point at.

Signed-off-by: Michael Rice <michael@michaelrice.org>