feat(index): Add appropriate struct tags for index generation.
The eventual goal of this work is to replace a bunch of code that
manually generates indexes with code that automatically figures
everything reasonable for statically defined types using reflection.

While we are at it, get rid of the client-side index tests, since
they will be entirely defined in terms of reflection on the server
side and are exercised pretty well there.