feat(modtimesha): Add GenerateStat()
Add a GenerateStat() method to ModTimeSha to allow passing in
pre-fetched stat() information instead of fetching it anew.  This is
intended to use when we are already doing lots of work on the
filesystem we are storing data on, and want to minimize the number
syscalls and potential disk seeks we get.