Add support for Windows batch file scripts within tasks
This commit fixes the issue of the agent not being able to find the
cmd.exe executable by using the COMSPEC environment variable.

This commit also uses the FromSlash function to ensure the path to the
task file uses the correct separator for the underlying OS.

This commit affects scripts within tasks for all supported operating
systems.